backend/requirements.txt
asgiref==3.8.1
cassandra-driver==3.29.1
click==8.1.7
cql==1.4.0
cqlsh==4.1.1
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
geomet==0.2.1.post1
linecache2==1.0.0
PyJWT==2.8.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
thrift==0.20.0
traceback2==1.4.0
typing_extensions==4.12.2
unittest2==1.1.0
